% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgen-package.R
\docType{package}
\name{rgen-package}
\alias{rgen}
\alias{rgen-package}
\title{This package provides header-only sampling routines.}
\description{
Provides popular sampling distributions C++ routines based in armadillo through a header file approach.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-rpkg/rgen}
  \item \url{https://r-pkg.thecoatlessprofessor.com/rgen/}
  \item Report bugs at \url{https://github.com/coatless-rpkg/rgen/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Steven Andrew Culpepper \email{sculpepp@illinois.edu} (\href{https://orcid.org/0000-0003-4226-6176}{ORCID}) [contributor]
}

}
