% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgen-package.R
\docType{package}
\name{rgen-package}
\alias{rgen}
\alias{rgen-package}
\title{rgen: Random Sampling Distribution C++ Routines for Armadillo}
\description{
Provides C++ armadillo headers for popular sampling distributions.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

Other contributors:
\itemize{
  \item Steven Culpepper \email{sculpepp@illinois.edu} [contributor]
}

}
