% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgen-package.R
\docType{package}
\name{rgen-package}
\alias{rgen}
\alias{rgen-package}
\title{This package provides header-only sampling routines.}
\description{
This package provides header-only sampling routines.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

Other contributors:
\itemize{
  \item Steven Culpepper \email{sculpepp@illinois.edu} [contributor]
}

}
